feat: Add new colors for secondary light and dark shades
This commit adds new color variables for the secondary light and dark shades in the `_variables.scss` file. It also adds corresponding TypeScript declarations in the `_exports.module.scss.d.ts` file. Additionally, the `Logo` component's SCSS file is updated to use the new colors for the `light` and `end` classes.